restruct: restructured project
