+++ plugin/src/main/java/net/yellowstrawberry/ecomonic/data/DatabaseType.java
