+++ plugin/src/main/resources/plugin.yml
