+++ settings.gradle
@@ -1,1 +1,7 @@
 rootProject.name = 'ecomonic'
+include 'api'
+include 'plugin'
+
+ext {
+    version = '0.0.1-SNAPSHOT'
+}
